Made a scraper that scrapes departments and their respective professors from the academic calendar of utm and returns a csv file.
